# fe/api-docs-admin.yaml
openapi: 3.0.1
info:
  title: Shop Admin API
  description: API documentation for Shop Admin
  version: v1
servers:
- url: http://localhost:8081
  description: Generated server url
paths:
...
  /api/admin/orders/{orderId}:
    get:
      tags:
      - order-admin-controller
      operationId: getOrder
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
...
components:
  schemas:
...
    OrderItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        productName:
          type: string
        variantId:
          type: integer
          format: int64
        size:
          type: string
        color:
          type: string
        unitPrice:
          type: number
        quantity:
          type: integer
          format: int32
        totalPrice:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        guestId:
          type: string
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        shippingAddress:
          type: string
        billingAddress:
          type: string
        status:
          type: string
          enum:
          - NEW
          - PENDING_PAYMENT
          - CONFIRMED
          - PROCESSING
          - PACKING
          - SHIPPING
          - DELIVERED
          - COMPLETED
          - CANCELLED
        paymentStatus:
          type: string
          enum:
          - PENDING
          - PAID
          - FAILED
          - REFUNDED
          - CANCELLED
        totalPrice:
          type: number
        shippingFee:
          type: number
        discountAmount:
          type: number
        finalAmount:
          type: number
        paymentMethod:
          type: string
        shippingMethod:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemResponse"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        paymentDate:
          type: string
          format: date-time
        estimatedDeliveryDate:
          type: string
          format: date-time
        actualDeliveryDate:
          type: string
          format: date-time
...